requirements.txt
streamlit==1.41.1
SQLAlchemy==2.0.36
pyodbc==5.1.0
python-dotenv==1.0.1
